Move coerce_types to types module.
